feat(ios): update Firebase to 7.11.0
