+++ ios/manifest
@@ -2,18 +2,18 @@
 # this is your module manifest and used by Titanium
 # during compilation, packaging, distribution, etc.
 #
-version: 5.0.0
+version: 6.0.0
 apiversion: 2
 architectures: arm64 x86_64
 description: titanium-firebase-cloud-messaging
 author: Hans Knoechel
 license: Apache 2
 mac: true
-copyright: Copyright (c) 2017 by Axway Appcelerator
+copyright: Copyright (c) 2017-present by Hans Knoechel
 
 # these should not be edited
 name: titanium-firebase-cloud-messaging
 moduleid: firebase.cloudmessaging
 guid: 36052f2b-8ddc-4177-b5a1-82c6b948f83d
 platform: iphone
-minsdk: 9.2.0.GA
+minsdk: 9.2.0
