--- ios/platform/FirebaseMessaging.xcframework/macos-x86_64/FirebaseMessaging.framework/Modules/module.modulemap
@@ -1,12 +1,0 @@
-framework module FirebaseMessaging {
-umbrella header "FirebaseMessaging.h"
-export *
-module * { export * }
-  link framework "AppKit"
-  link framework "CoreTelephony"
-  link framework "Foundation"
-  link framework "Security"
-  link framework "SystemConfiguration"
-  link "sqlite3"
-  link "z"
-}
