+++ ios/titanium.xcconfig
@@ -4,7 +4,7 @@
 // OF YOUR TITANIUM SDK YOU'RE BUILDING FOR
 //
 //
-TITANIUM_SDK_VERSION = 9.3.1.GA
+TITANIUM_SDK_VERSION = 9.3.2.GA
 
 //
 // THESE SHOULD BE OK GENERALLY AS-IS
